[refactor] salePostImage: salepost 패키지에서 독립 패키지로 이관
- 다른 도메인과 마찬가지로, 각 도메인 분리
